Create usage report service
Move methods from organization service
