+++ lms/tasks/organization.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from lms.services import OrganizationService
+from lms.services import OrganizationUsageReportService
 from lms.tasks.celery import app
 
 
@@ -10,7 +10,7 @@ def generate_usage_report(
 ) -> None:
     with app.request_context() as request:
         with request.tm:
-            request.find_service(OrganizationService).generate_usage_report(
+            request.find_service(OrganizationUsageReportService).generate_usage_report(
                 organization_id,
                 tag,
                 datetime.fromisoformat(since),
