+++ tests/unit/lms/services/organization_usage_report_test.py
@@ -1,0 +1,224 @@
+from datetime import datetime, timedelta
+from unittest.mock import patch, sentinel
+
+import pytest
+from h_matchers import Any
+
+from lms.services.h_api import HAPI
+from lms.services.organization_usage_report import (
+    OrganizationUsageReportService,
+    UsageReportRow,
+    service_factory,
+)
+from tests import factories
+
+
+class TestOrganizationUsageReportService:
+    def test_generate_usage_report(
+        self, svc, org_with_parent, usage_report, organization_service
+    ):
+        organization_service.get_by_id.return_value = org_with_parent
+        usage_report.return_value = [
+            UsageReportRow(
+                name="<STUDENT>",
+                email="<STUDENT>",
+                h_userid=sentinel.h_userid,
+                course_name=sentinel.lms_name,
+                course_created="2020-01-01",
+                authority_provided_id=sentinel.authority_provided_id,
+            ),
+            UsageReportRow(
+                name="<STUDENT>",
+                email="<STUDENT>",
+                h_userid=sentinel.h_userid,
+                course_name=sentinel.lms_name,
+                course_created="2020-01-01",
+                authority_provided_id=sentinel.authority_provided_id,
+            ),
+        ]
+
+        report = svc.generate_usage_report(
+            org_with_parent.id, "test", "2020-01-01", "2020-02-02"
+        )
+
+        usage_report.assert_called_once_with(
+            org_with_parent, "2020-01-01", "2020-02-02"
+        )
+
+        assert report.organization == org_with_parent
+        assert report.unique_users == 1
+        assert report.since == "2020-01-01"
+        assert report.until == "2020-02-02"
+        assert len(report.report) == 2
+
+    def test_generate_usage_report_existing_report(
+        self, svc, org_with_parent, organization_service
+    ):
+        organization_service.get_by_id.return_value = org_with_parent
+
+        report = factories.OrganizationUsageReport(
+            organization=org_with_parent,
+            key=f"{org_with_parent.public_id}-test-2020-01-01-2020-02-02",
+            tag="test",
+        )
+
+        assert (
+            svc.generate_usage_report(
+                org_with_parent.id, "test", "2020-01-01", "2020-02-02"
+            )
+            == report
+        )
+
+    def test_usage_report(self, svc, org_with_parent, h_api, organization_service):  # pylint:disable=too-many-locals
+        since = datetime(2023, 1, 1)
+        until = datetime(2023, 12, 31)
+
+        ai_root_org = factories.ApplicationInstance(organization=org_with_parent.parent)
+        ai_child_org = factories.ApplicationInstance(organization=org_with_parent)
+        course_root = factories.Course(
+            application_instance=ai_root_org,
+            created=since + timedelta(days=1),
+        )
+        section = factories.CanvasSection(
+            application_instance=ai_root_org,
+            parent=course_root,
+            created=since + timedelta(days=1),
+        )
+        course_child = factories.Course(
+            application_instance=ai_child_org, created=since + timedelta(days=1)
+        )
+        # Course created after the until date
+        factories.Course(
+            application_instance=ai_child_org,
+            created=until + timedelta(days=1),
+        )
+        # Annotations in one section and in the other course
+        h_api.get_groups.return_value = [
+            HAPI.HAPIGroup(authority_provided_id=group.authority_provided_id)
+            for group in [section, course_child]
+        ]
+        # Users that belong to the course
+        user_1 = factories.User(display_name="NAME", email="EMAIL")
+        user_2 = factories.User()
+        factories.GroupingMembership(
+            user=user_1, grouping=course_child, created=since + timedelta(days=1)
+        )
+        factories.GroupingMembership(
+            user=user_2, grouping=course_root, created=since + timedelta(days=1)
+        )
+        organization_service.get_hierarchy_ids.return_value = [
+            org_with_parent.parent.id,
+            org_with_parent.id,
+        ]
+
+        report = svc.usage_report(org_with_parent.parent, since, until)
+
+        h_api.get_groups.assert_called_once_with(
+            Any.list.containing(
+                [
+                    course_root.authority_provided_id,
+                    course_child.authority_provided_id,
+                    section.authority_provided_id,
+                ]
+            ),
+            since,
+            until,
+        )
+
+        # We expect to get both users belonging to each course
+        expected = [
+            UsageReportRow(
+                name=user_1.display_name,
+                email=user_1.email,
+                h_userid=user_1.h_userid,
+                course_name=course_child.lms_name,
+                course_created=course_child.created.date().isoformat(),
+                authority_provided_id=course_child.authority_provided_id,
+            ),
+            UsageReportRow(
+                name="<STUDENT>",
+                email="<STUDENT>",
+                h_userid=user_2.h_userid,
+                course_name=course_root.lms_name,
+                course_created=course_root.created.date().isoformat(),
+                authority_provided_id=course_root.authority_provided_id,
+            ),
+        ]
+        assert report == Any.list.containing(expected)
+
+    def test_usage_report_with_no_courses(self, svc, org_with_parent):
+        since = datetime(2023, 1, 1)
+        until = datetime(2023, 12, 31)
+
+        with pytest.raises(ValueError) as error:
+            svc.usage_report(org_with_parent.parent, since, until)
+
+        assert "no courses found" in str(error.value).lower()
+
+    def test_usage_report_with_no_activity(
+        self, svc, org_with_parent, h_api, organization_service
+    ):
+        organization_service.get_hierarchy_ids.return_value = [
+            org_with_parent.parent.id,
+            org_with_parent.id,
+        ]
+
+        since = datetime(2023, 1, 1)
+        until = datetime(2023, 12, 31)
+
+        ai_root_org = factories.ApplicationInstance(organization=org_with_parent.parent)
+        factories.Course(
+            application_instance=ai_root_org, created=since + timedelta(days=1)
+        )
+        h_api.get_groups.return_value = []
+
+        with pytest.raises(ValueError) as error:
+            svc.usage_report(org_with_parent.parent, since, until)
+
+        assert "no courses with activity" in str(error.value).lower()
+
+    @pytest.fixture
+    def org_with_parent(self, db_session):
+        org_with_parent = factories.Organization.create(
+            parent=factories.Organization.create()
+        )
+        # Flush to ensure public ids are generated
+        db_session.flush()
+        return org_with_parent
+
+    @pytest.fixture
+    def svc(self, db_session, h_api, organization_service):
+        return OrganizationUsageReportService(
+            db_session=db_session,
+            h_api=h_api,
+            organization_service=organization_service,
+        )
+
+    @pytest.fixture
+    def usage_report(self, svc):
+        with patch.object(svc, "usage_report") as usage_report:
+            yield usage_report
+
+
+class TestServiceFactory:
+    def test_it(
+        self,
+        pyramid_request,
+        OrganizationUsageReportService,
+        h_api,
+        organization_service,
+    ):
+        svc = service_factory(sentinel.context, pyramid_request)
+
+        OrganizationUsageReportService.assert_called_once_with(
+            db_session=pyramid_request.db,
+            h_api=h_api,
+            organization_service=organization_service,
+        )
+        assert svc == OrganizationUsageReportService.return_value
+
+    @pytest.fixture
+    def OrganizationUsageReportService(self, patch):
+        return patch(
+            "lms.services.organization_usage_report.OrganizationUsageReportService"
+        )
